requirements.txt
gradio
chromadb
sentence-transformers
python-dotenv
google-generativeai
rapidfuzz